requirements.txt
boto3==1.34.34
botocore==1.34.34
aws-sso-lib==1.14.0
PyYAML==6.0.1